requirements.txt
autogen-agentchat
autogen-ext[ollama,openai]
python-dotenv